🐛 Set the length of value buffer
This was not done before on the wasi guest side which meant
that even though it received data it treated it as size 0
